refactor(lib): generalize news-routes into routes and add resolveLocale
Rename src/lib/news-routes.ts to src/lib/routes.ts since its path
builders also serve the papers section, and add shared helpers
(withTrailingSlash, toLangParam, getLangStaticPaths,
buildNewsOgImagePath). Add resolveLocale() to i18n to replace the
repeated `isLocale(x) ? x : defaultLocale` pattern.
