+++ src/pages/404.astro
@@ -3,7 +3,7 @@ import NewsLayout from "@/components/news/NewsLayout.astro";
 import NewsNotFoundState from "@/components/news/NewsNotFoundState.astro";
 import Layout from "@/layouts/main.astro";
 import { getLocaleFromPath, t } from "@/lib/i18n";
-import { buildSectionIndexPath } from "@/lib/news-routes";
+import { buildSectionIndexPath } from "@/lib/routes";
 
 const locale = getLocaleFromPath(Astro.url.pathname);
 const basePath = import.meta.env.BASE_URL;
