+++ src/pages/news/[slug].astro
@@ -9,7 +9,7 @@ import {
   getNewsArticleState,
   type NewsEntry,
 } from "@/lib/news";
-import { buildSectionIndexPath } from "@/lib/news-routes";
+import { buildSectionIndexPath } from "@/lib/routes";
 import { getCollection, render } from "astro:content";
 
 export async function getStaticPaths() {
